+++ src/OptionManager.sol
@@ -8,7 +8,7 @@ import {IOptionToken} from "@lyra-protocol/contracts/interfaces/IOptionToken.sol
 contract TraderExample is LyraAdapter {
     constructor() LyraAdapter() {}
 
-    uint[] public activePositionIds;
+    uint256[] public activePositionIds;
 
     function initAdapter(address _lyraRegistry, address _optionMarket, address _curveSwap, address _feeCounter)
         external
@@ -17,26 +17,26 @@ contract TraderExample is LyraAdapter {
         setLyraAddresses(_lyraRegistry, _optionMarket, _curveSwap, _feeCounter);
     }
 
-    function modifyLyraPosition(uint positionId, uint amount, uint collateral) external onlyOwner{
-      LyraAdapter.OptionPosition[] memory positions = _getPositions(_singletonArray(positionId)); // must first convert number into a static array
-      // Position position = _getPositions(_singletonArray(positionId)); // must first convert number into a static array
+    function modifyLyraPosition(uint256 positionId, uint256 amount, uint256 collateral) external onlyOwner {
+        LyraAdapter.OptionPosition[] memory positions = _getPositions(_singletonArray(positionId)); // must first convert number into a static array
+        // Position position = _getPositions(_singletonArray(positionId)); // must first convert number into a static array
 
-      LyraAdapter.OptionPosition memory position = positions[0];
+        LyraAdapter.OptionPosition memory position = positions[0];
 
-      TradeInputParameters memory tradeParams = TradeInputParameters({
-        strikeId: position.strikeId,
-        positionId: position.positionId,
-        iterations: 3,
-        optionType: position.optionType,
-        amount: amount, // closing 100%
-        setCollateralTo: collateral, // increase collateral by addCollatAmount
-        minTotalCost: 0,
-        maxTotalCost: type(uint).max, // assume we are ok with any premium amount
-        rewardRecipient: address(0)
-      });
+        TradeInputParameters memory tradeParams = TradeInputParameters({
+            strikeId: position.strikeId,
+            positionId: position.positionId,
+            iterations: 3,
+            optionType: position.optionType,
+            amount: amount, // closing 100%
+            setCollateralTo: collateral, // increase collateral by addCollatAmount
+            minTotalCost: 0,
+            maxTotalCost: type(uint256).max, // assume we are ok with any premium amount
+            rewardRecipient: address(0)
+        });
 
-      // built-in LyraAdapter.sol functions
-      _closeOrForceClosePosition(tradeParams);
+        // built-in LyraAdapter.sol functions
+        _closeOrForceClosePosition(tradeParams);
     }
 
     function _singletonArray(uint256 val) public view returns (uint256[] memory arr) {
